requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar
pytest
pytest-django
pytest-cov
factory-boy
faker
ipython
django-silk